requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
protobuf==3.20.*
